feat(Turnips): add Turnips chart part 7
* Improve #59 Provide a basic Chart for Turnips Week prices
